requirements.txt
numpy==1.17.2
opencv_python==4.2.0.34
Pillow==8.0.0
